TASK: Set output to the document identifier for correct deduplication
For suggestions, two hits are deduplicated when they prouce the same output. By default it make sense to deduplicate hits by the linked document, so the document identifier is used. This also reduces the index document size a lot.

Also improved documentation and changed the method name as it is not a config that is built.